refactor(test): split test_codeagent_config.py into focused modules
Split 682-line test file into 5 focused modules:
- test_agent_preset_resolution.py (240 lines) - preset resolution logic
- test_sync_models_json.py (150 lines) - models.json sync behavior
- test_models_json_path.py (73 lines) - path override tests
- test_env_var_override.py (219 lines) - env var override tests
- test_find_missing_backend_commands.py (32 lines) - backend availability checks

Follows MANY SMALL FILES principle for better maintainability.
